% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-bed.R
\name{write_bed}
\alias{write_bed}
\alias{write_bed_graph}
\alias{write_narrowpeaks}
\title{Write a BED or BEDGraph file}
\usage{
write_bed(x, file, index = FALSE)

write_bed_graph(x, file, index = FALSE)

write_narrowpeaks(x, file)
}
\arguments{
\item{x}{A GRanges object}

\item{file}{File name, URL or connection specifying a file to write x to.
Compressed files with extensions such as '.gz' are handled
automatically. If you want to index the file with tabix use the
\code{index} argument.}

\item{index}{Compress and index the output file
with bgzf and tabix (default = FALSE). Note that tabix indexing will sort the
data by chromosome and start.}
}
\value{
The write functions return a BED(Graph)File invisibly
}
\description{
This is a lightweight wrapper to the export family
of functions defined in \pkg{rtracklayer}.
}
\seealso{
\code{\link[rtracklayer:BEDFile]{rtracklayer::BEDFile()}}
}
