% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-gff.R
\name{write_gff}
\alias{write_gff}
\alias{write_gff1}
\alias{write_gff2}
\alias{write_gff3}
\title{Write a GFF(123) file}
\usage{
write_gff(x, file, index = FALSE)

write_gff1(x, file, index = FALSE)

write_gff2(x, file, index = FALSE)

write_gff3(x, file, index = FALSE)
}
\arguments{
\item{x}{A GRanges object}

\item{file}{Path or connection to write to}

\item{index}{If TRUE the output file will be compressed and indexed using bgzf and
tabix.}
}
\value{
The write function returns a GFFFile object invisibly
}
\description{
This is a lightweight wrapper to the export family
of functions defined in \pkg{rtracklayer}.
}
\seealso{
\code{\link[rtracklayer:GFFFile]{rtracklayer::GFFFile()}}
}
