% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-wig.R
\name{write_wig}
\alias{write_wig}
\title{Write a WIG file}
\usage{
write_wig(x, file)
}
\arguments{
\item{x}{A GRanges object}

\item{file}{File name, URL or connection specifying a file to write x to.
Compressed files with extensions such as '.gz' are handled
automatically.}
}
\value{
The write function returns a WIGFile invisibly.
}
\description{
Write a WIG file
}
\seealso{
\code{rtracklayer::\link[rtracklayer:WIGFile-class]{WIGFile()}}
}
