% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-join-mcols.R
\name{join_mcols_left}
\alias{join_mcols_left}
\title{Join data by metadata columns}
\usage{
join_mcols_left(x, y, ...)
}
\arguments{
\item{x}{Object representing ranges, with metadata
columns containing variables for matching}

\item{y}{A table of data, DataFrame, data.frame or tibble}

\item{...}{arguments passed to \code{dplyr::left_join}}
}
\value{
Object representing ranges, with new metadata columns
}
\description{
Utilizing DFplyr which enables dplyr verbs for DataFrame
objects, these functions enable joins based on metadata
of \code{x} and data in \code{y}.
}
\examples{

x <- as_granges(data.frame(
  seqnames = "chr1", start = 1:5, width=10,
  id=c(1:4,4), foo=letters[1:5]
))
y <- DataFrame(id=c(1:2,2,4), bar=letters[6:9])
 
# metadata join
# join_mcols_left(x, y, by="id")

}
