% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-add-distance.R
\name{make_add_nearest_distance}
\alias{make_add_nearest_distance}
\title{Macro for building add_nearest_distance_* family of functions
Change this to edit defaults for all functions}
\usage{
make_add_nearest_distance(fun)
}
\description{
Macro for building add_nearest_distance_* family of functions
Change this to edit defaults for all functions
}
