% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-eval.R
\name{overscope_ranges}
\alias{overscope_ranges}
\title{Create an overscoped environment from a Ranges object}
\usage{
overscope_ranges(x, envir = parent.frame())
}
\arguments{
\item{x}{a Ranges object}

\item{envir}{the environment to place the Ranges in (default = \code{parent.frame()})}
}
\description{
Create an overscoped environment from a Ranges object
}
\details{
This is the backend for non-standard evaluation in \code{plyranges}.
}
\seealso{
\code{\link[rlang:new_data_mask]{rlang::new_data_mask()}}, \code{\link[rlang:eval_tidy]{rlang::eval_tidy()}}
}
