% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-arrange.R
\name{arrange.Ranges}
\alias{arrange.Ranges}
\title{Sort a Ranges object}
\usage{
\method{arrange}{Ranges}(.data, ...)
}
\arguments{
\item{.data}{A Ranges object.}

\item{...}{Comma seperated list of variable names.}
}
\value{
A sorted Ranges object
}
\description{
Sort a Ranges object
}
\examples{
rng <- as_iranges(data.frame(start = 1:10, width = 10:1))
rng <- mutate(rng, score = runif(10))
arrange(rng, score)
# you can also use dplyr::desc to arrange by descending order
}
