% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-Operator.R
\docType{class}
\name{FileOperator-class}
\alias{FileOperator-class}
\alias{BamFileOperator-class}
\title{An abstract class to represent operations performed over a file}
\description{
An abstract class to represent operations performed over a file
}
\details{
This class is used internally by DeferredGenomicRanges objects.
Currently, this class is only implemented for bam files (as a
BamFileOperator) but will eventually be extended to the other avaialable
readers.
}
