% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io-deferred-Classes.R
\docType{class}
\name{GRangesDeferred-class}
\alias{GRangesDeferred-class}
\alias{GRangesDeferred}
\alias{has_cache_contents}
\alias{get_cache}
\alias{get_operation}
\title{GRangesDeferred Class}
\usage{
GRangesDeferred(cache = GRanges(), operation)

has_cache_contents(.data)

get_cache(.data)

get_operation(.data)
}
\arguments{
\item{cache}{a GRanges object}

\item{operation}{an environment containing operations to perform
to generate the cache.}

\item{.data}{a GRangesDeferred object}
}
\value{
a GRangesDeferred object.
}
\description{
Enables deferred reading of BAM files by caching
after performing file operations.
}
\details{
To see what is loaded in the cache use
\code{get_cache} and to see the file operations being performed
use \code{get_operation}. To check whether the cache is full
use \code{has_cache_contents}.
By default the \code{show} method for a GRangesDeferred object
shows what's in the cache.
}
\seealso{
read_bam
}
