% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-pull.R
\name{pull.Ranges}
\alias{pull.Ranges}
\title{Extract a single column of (meta)data from a Ranges object by name or position}
\usage{
\method{pull}{Ranges}(.data, var = -1, name = NULL)
}
\arguments{
\item{.data}{a \code{Ranges} object}

\item{var}{One column name, either from the core GRanges object or its metadata. If empty will return the last column in mdata.}

\item{name}{An optional parameter that specifies the column to be used as names for a named vector. Specified in a similar manner as var.}

\item{...}{Not used in this method.}
}
\value{
A vector the same size as the number of ranges in .data.
}
\description{
Extract a single column of (meta)data from a Ranges object by name or position
}
\examples{
df <- data.frame(start = 1:10, width = 5,  seqnames = "seq1",
strand = sample(c("+", "-", "*"), 10, replace = TRUE), gc = runif(10), counts = rpois(10, 2))
rng <- as_granges(df)
pull(rng, gc)
pull(rng, counts)
pull(rng, start)
}
\seealso{
\code{\link[dplyr:pull]{dplyr::pull()}}
}
