% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-arithmetic-setters.R
\name{set_width}
\alias{set_width}
\alias{set_start}
\alias{set_end}
\alias{set_seqnames}
\alias{set_strand}
\title{Functional setters for Ranges objects}
\usage{
set_width(x, width)

set_start(x, start = 0L)

set_end(x, end = 0L)

set_seqnames(x, seqnames)

set_strand(x, strand)
}
\arguments{
\item{x}{a Ranges object}

\item{width}{integer amount to modify width by}

\item{start}{integer amount to modify start by}

\item{end}{integer amount to modify end by}

\item{seqnames}{update seqnames column}

\item{strand}{update strand column}
}
\value{
a Ranges object
}
\description{
Functional setters for Ranges objects
}
\details{
These methods are used internally in \code{mutate()} to modify
core columns in Ranges objects.
}
