% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-slice.R
\name{slice.Ranges}
\alias{slice.Ranges}
\title{Choose rows in a Ranges object by postion}
\usage{
\method{slice}{Ranges}(.data, ..., .preserve = FALSE)
}
\arguments{
\item{.data}{a Ranges object}

\item{...}{integer row values giving the position}

\item{.preserve}{not yet implemented}
}
\value{
a Ranges object
}
\description{
Choose rows in a Ranges object by postion
}
\examples{
set.seed(100)
df <- data.frame(start = 1:10,
                 width = 5,
                 seqnames = "seq1",
                 strand = sample(c("+", "-", "*"), 10, replace = TRUE),
                 gc = runif(10))

rng <- as_granges(df)

dplyr::slice(rng, n())
dplyr::slice(rng, 3:n())
dplyr::slice(rng, which.max(gc))

# grouping acts on each subset of the data
by_strand <- rng \%>\%
  group_by(strand) 

by_strand \%>\% 
  dplyr::slice(n())

by_strand \%>\% 
 dplyr::slice(which.min(gc), which.max(gc))


}
