% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-arithmetic-width.R
\name{set_width}
\alias{set_width}
\title{Set the width of a genomic interval}
\usage{
set_width(x, width)
}
\arguments{
\item{x}{a Ranges object, to fix by either the start, end or center
of an interval use \code{anchor_start(x)}, \code{anchor_end(x)},
\code{anchor_center(x)}. To fix by start for negative strand and
end by positive strand use \code{anchor_3p(x)}. To fix by start for positive
strand and end by negative strand use \code{anchor_end(x)}.}

\item{width}{the amount to alter the width of a Ranges object by. Either an
integer vector of length 1 or an integer vector the same length as x.}
}
\value{
a Ranges object with modified width
}
\description{
Set the width of a genomic interval
}
\details{
By default unstranded features are treated as
positive strand features.
}
\examples{
rng <- as_iranges(data.frame(start=c(2:-1, 13:15), width=c(0:3, 2:0)))
set_width(rng, width = 10)
set_width(anchor_start(rng), 10)
set_width(anchor_end(rng), 10)
set_width(anchor_center(rng), 10)
grng <- as_granges(data.frame(seqnames = "chr1",
                         strand = c("+", "-", "-", "+", "+", "-", "+"),
                         start=c(2:-1, 13:15),
                         width=c(0:3, 2:0)))
set_width(anchor_3p(grng), 10)
set_width(anchor_5p(grng), 10)
}
\seealso{
\code{\link[IRanges]{resize}}
}
