% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-groups-utils.R
\name{split_groups}
\alias{split_groups}
\title{Split a Grouped Ranges object into a RangesList}
\usage{
split_groups(.data_grouped, populate_mcols = FALSE, drop = TRUE)
}
\arguments{
\item{.data_grouped}{A grouped ranges object}

\item{populate_mcols}{add grouping to metadata slot of RangesList}

\item{drop}{drop unused group levels}
}
\value{
a RangesList object
}
\description{
Split a Grouped Ranges object into a RangesList
}
