% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reexports.R
\name{!!!}
\alias{!!!}
\alias{!!}
\alias{\%>\%}
\alias{filter}
\alias{group_by}
\alias{groups}
\alias{ungroup}
\alias{group_vars}
\alias{mutate}
\alias{select}
\alias{summarise}
\alias{summarize}
\alias{arrange}
\alias{slice}
\title{Objects exported from other packages.}
\usage{
`!!!`(x)

`!!`(x)

lhs \%>\% rhs

filter(.data, ..., .by = NULL, .preserve = FALSE)

group_by(.data, ..., .add = FALSE, .drop = group_by_drop_default(.data))

groups(x)

ungroup(x, ...)

group_vars(x)

mutate(.data, ...)

select(.data, ...)

summarise(.data, ..., .by = NULL, .groups = NULL)

summarize(.data, ..., .by = NULL, .groups = NULL)

arrange(.data, ..., .by_group = FALSE)

slice(.data, ..., .by = NULL, .preserve = FALSE)
}
\description{
Objects exported from other packages.
}
