% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-mutate.R
\name{mutate.GenomicRanges}
\alias{mutate.GenomicRanges}
\alias{mutate.Ranges}
\alias{mutate.GRangesGrouped}
\alias{mutate.IRangesGrouped}
\title{Modify a Ranges object}
\usage{
\method{mutate}{GenomicRanges}(.data, ...)

\method{mutate}{Ranges}(.data, ...)

\method{mutate}{GRangesGrouped}(.data, ...)

\method{mutate}{IRangesGrouped}(.data, ...)
}
\arguments{
\item{.data}{a \code{Ranges} object}

\item{...}{Pairs of name-value expressions. The name-value pairs can either
create new metadata columns or modify existing ones.}
}
\value{
a Ranges object
}
\description{
Modify a Ranges object
}
\examples{
df <- data.frame(start = 1:10,
                 width = 5,
                 seqnames = "seq1",
                 strand = sample(c("+", "-", "*"), 10, replace = TRUE),
                 gc = runif(10))
rng <- as_granges(df)

# mutate adds new columns
rng \%>\%
    mutate(avg_gc = mean(gc), row_id = 1:n())
# can also compute on newly created columns
rng \%>\%
    mutate(score = gc * width, score2 = score + 1)
# group by partitions the data and computes within each group
rng \%>\%
    group_by(strand) \%>\%
    mutate(avg_gc = mean(gc), row_id = 1:n())

}
