% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-DeferredGenomicRanges.R
\docType{class}
\name{DeferredGenomicRanges-class}
\alias{DeferredGenomicRanges-class}
\title{DeferredGenomiRanges objects}
\description{
Enables deferred reading of files (currently only BAM files) by
caching results after a plyranges verb is called.
}
\section{Slots}{

\describe{
\item{\code{delegate}}{a GenomicRanges object to be cached}

\item{\code{ops}}{A FileOperator object}
}}

\seealso{
read_bam
}
