% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-summarize.R
\name{summarise.GenomicRanges}
\alias{summarise.GenomicRanges}
\alias{summarise.Ranges}
\alias{summarise.GRangesGrouped}
\alias{summarise.IRangesGrouped}
\title{Aggregate a Ranges object}
\usage{
\method{summarise}{GenomicRanges}(.data, ...)

\method{summarise}{Ranges}(.data, ...)

\method{summarise}{GRangesGrouped}(.data, ...)

\method{summarise}{IRangesGrouped}(.data, ...)
}
\arguments{
\item{.data}{a Ranges object}

\item{...}{Name-value pairs of summary functions.}
}
\value{
a \link[S4Vectors]{DataFrame}
}
\description{
Aggregate a Ranges object
}
\examples{
df <- data.frame(start = 1:10, width = 5,  seqnames = "seq1",
strand = sample(c("+", "-", "*"), 10, replace = TRUE), gc = runif(10))
rng <- as_granges(df)
rng \%>\% summarise(gc = mean(gc))
rng \%>\% group_by(strand) \%>\% summarise(gc = mean(gc))
}
