% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dplyr-utils.R, R/rlang-utils.R,
%   R/tidyr-unnest.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{\%>\%}
\alias{filter}
\alias{group_by}
\alias{groups}
\alias{ungroup}
\alias{group_vars}
\alias{mutate}
\alias{select}
\alias{summarise}
\alias{summarize}
\alias{arrange}
\alias{enquo}
\alias{UQS}
\alias{UQ}
\alias{unnest}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{arrange}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{enquo}}, \code{\link[rlang]{UQS}}, \code{\link[rlang]{UQ}}}

  \item{tidyr}{\code{\link[tidyr]{unnest}}}
}}

