% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ranges-eval.R, R/reexports.R
\docType{import}
\name{tbl_vars.Ranges}
\alias{tbl_vars.Ranges}
\alias{tbl_vars.GenomicRanges}
\alias{tidyverse-reexports}
\alias{!!!}
\alias{reexports}
\alias{!!}
\alias{\%>\%}
\alias{filter}
\alias{group_by}
\alias{groups}
\alias{ungroup}
\alias{group_vars}
\alias{mutate}
\alias{select}
\alias{summarise}
\alias{summarize}
\alias{arrange}
\alias{slice}
\title{Re-exports
Rexported functions from the tidyverse}
\usage{
\method{tbl_vars}{Ranges}(x)

\method{tbl_vars}{GenomicRanges}(x)
}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{dplyr}{\code{\link[dplyr]{filter}}, \code{\link[dplyr]{group_by}}, \code{\link[dplyr]{groups}}, \code{\link[dplyr]{ungroup}}, \code{\link[dplyr]{group_vars}}, \code{\link[dplyr]{mutate}}, \code{\link[dplyr]{select}}, \code{\link[dplyr]{summarise}}, \code{\link[dplyr]{summarize}}, \code{\link[dplyr]{arrange}}, \code{\link[dplyr]{slice}}}

  \item{magrittr}{\code{\link[magrittr]{\%>\%}}}

  \item{rlang}{\code{\link[rlang]{!!!}}, \code{\link[rlang]{!!}}}
}}

